:recycle: split css files - issue#4
