+++ src/components/Links.js
@@ -1,5 +1,5 @@
 import { Inter } from '@next/font/google'
-import styles from '@/styles/Home.module.css'
+import styles from '@/components/Links.module.css'
 
 const inter = Inter({ subsets: ['latin'] })
