+++ src/components/Main.module.css
@@ -1,0 +1,13 @@
+.main {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: center;
+  padding: 6rem;
+  min-height: 100vh;
+}
+
+.code {
+  font-weight: 700;
+  font-family: var(--font-mono);
+}
