+++ src/components/NextLogo.js
@@ -1,6 +1,6 @@
 import Image from 'next/image'
 import { Inter } from '@next/font/google'
-import styles from '@/styles/Home.module.css'
+import styles from '@/components/NextLogo.module.css'
 
 const inter = Inter({ subsets: ['latin'] })
